+++ src/main/java/tech/jazz/apianalisecredito/presentation/dto/response/ClientAnalysisResponse.java
@@ -10,7 +10,7 @@ public record ClientAnalysisResponse(
         BigDecimal approvedLimit,
         BigDecimal requestedAmount,
         BigDecimal withdraw,
-        Integer annualInterest,
+        Float annualInterest,
         String clientId,
         LocalDateTime date
 ) {
